solved day16 part 2 in C

+++ day16/src/node.h
@@ -1,6 +1,8 @@
 #ifndef AOC16_NODE_H
 #define AOC16_NODE_H
 
+#include <stdint.h>
+
 enum NodeType {
   LiteralValue,
   Operator,
@@ -16,7 +18,7 @@ typedef struct Node {
   int packetVersion;
   union {
     struct {
-      int value;
+      uint64_t value;
     } literal;
 
     struct {
